+++ src/main/kotlin/com/bot/telegram/gamemaster/messages/BotMessage.kt
@@ -1,3 +1,3 @@
 package com.bot.telegram.gamemaster.messages
 
-class BotMessage(val chat_id: Int, val text: String?)
+class BotMessage(val chat_id: Long, val text: String?)
