+++ src/main/kotlin/com/bot/telegram/gamemaster/messages/Chat.kt
@@ -1,4 +1,4 @@
 package com.bot.telegram.gamemaster.messages
 
 
-data class Chat(val id: Int, val bot: Boolean, val firstName: String?, val lastName: String?, val type: String?)
+data class Chat(val id: Long, val bot: Boolean, val firstName: String?, val lastName: String?, val type: String?)
